requirements.txt
mysql-connector-python>2.0
requests
